backend/requirements.txt
fastapi
uvicorn
pydantic
asyncpg
sqlalchemy
databases
python-dotenv
psycopg2